+++ changelog.md
@@ -18,6 +18,7 @@ Codion Change Log
 - EntityTablePanel, TableControl.EDIT_SELECTED renamed EDIT_VALUE and now associated with an action for editing the selected items.
 - EntityTablePanel.Config.editAttributeSelection() and includeEditValueControl() added.
 - EntityTablePanel.TableControl.DELETE_SELECTED renamed DELETE, related renaming.
+- EntityTablePanel, edit value control added to toolbar.
 
 ## 0.17.41
 ### is.codion.common.core